add pipeline readme; small updates
