+++ .gitignore
@@ -177,3 +177,4 @@ scratch.ipynb
 
 # default output directory for nl2code pipeline 
 nl2code-artifacts/
+pipeline-artifacts/
